ported from double to float

+++ src/main/scala/events.scala
@@ -7,7 +7,7 @@ abstract class NoiseEditorEvent(source:Publisher) extends Event {
 	//println("Event: " + this)
 }
 
-case class NodeValueChanged(source:Publisher, node:Node, slider:String, value:Float) extends NoiseEditorEvent(source)
+case class NodeValueChanged(source:Publisher, node:Node, slider:String, value:Double) extends NoiseEditorEvent(source)
 case class NodeChanged(source:Publisher, node:Node) extends NoiseEditorEvent(source)
 case class HitConnector(source:Publisher, connector:Connector) extends NoiseEditorEvent(source)
 case class NodeMoved(source:Publisher, node:Node) extends NoiseEditorEvent(source)
